Change ErrorAction from SilentlyContinue to Ignore
If you use `SilentlyContinue` the error gets added to `$global:error`, which is not what you would expect, as you don't intend to inpect the error later, and causes side effects. Simply ignore the error.

I found this when I added this module and started seeing errors on PowerShell startup, which should be clean.